+++ src/server/server_main_tab_plots.R
@@ -13,6 +13,33 @@ observeEvent(input$cancelPlot, {
   killProcessTree(currentPlotPid(), warnFail=TRUE)
 }, priority=2000, ignoreInit=TRUE)
 
+# Management of plot progress bar
+plotProgressFile <- reactiveVal(NULL)
+plotProgressStatus <- reactiveVal(function() NULL)
+plotProgressBar <- reactiveVal(NULL)
+observeEvent(plotProgressFile(), {
+  plotProgressStatus(reactiveFileReader(
+    500, session, isolate(plotProgressFile()), readPlotProgressFile
+  ))
+})
+observeEvent(plotProgressStatus()(), {
+  pProgress <- plotProgressStatus()()
+  req(pProgress$current, pProgress$total)
+  removeNotification(plotStartedNotifId())
+  progress <- plotProgressBar()
+  if(is.null(progress)) {
+    progress <- shiny::Progress$new(max=pProgress$total)
+    plotProgressBar(progress)
+  }
+  progress$set(
+    # Subtract 1 from value as progress is updated when the process begins
+    value=pProgress$current-1,
+    message=sprintf(
+      "Plot: Querying data file %s of %s", pProgress$current, pProgress$total
+    )
+  )
+})
+
 readyPlot <- reactiveVal(NULL)
 observeEvent(input$doPlot, {
   # Make sure a plot cannot be requested if another is being produced.
@@ -60,6 +87,13 @@ observeEvent(input$doPlot, {
   shinyjs::enable("cancelPlot")
   shinyjs::show("cancelPlot")
 
+  plotStartedNotifId(showNotification(
+    "Processing plot request...", type="message", duration=NULL
+  ))
+
+  # Trigger creation of progress bar
+  plotProgressFile(tempfile(pattern="plotProgress"))
+
   # The plot tab does not keep the spinner running if the plot
   # is NULL, but the plotly tab does. Using this to keep the
   # spinner running while the plot is being prepared.
@@ -69,16 +103,13 @@ observeEvent(input$doPlot, {
   }
   shinyjs::hide(selector="#mainArea li a[data-value=plotTab]")
 
-  plotStartedNotifId(showNotification(
-    "Gathering data for plot...", type="message", duration=NULL
-  ))
-
   # Prepare plot asyncronously
   newFutPlotAndOutput <- futureCall(
     FUN=preparePlotsCapturingOutput,
     args=list(
       plotter=plotter, plotRequest=plotRequest, db=db,
-      interactive=isTRUE(obsmonConfig$general$plotsEnableInteractivity)
+      interactive=isTRUE(obsmonConfig$general$plotsEnableInteractivity),
+      progressFile=plotProgressFile()
     )
   )
   plotPID <- newFutPlotAndOutput$job$pid
@@ -141,7 +172,13 @@ observeEvent(input$doPlot, {
   )
   plotCleanup <- finally(newFutPlotAndOutput, function() {
     currentPlotPid(-1)
+    # Reset items related to plot progress bar
     removeNotification(plotStartedNotifId())
+    unlink(plotProgressFile())
+    if(!is.null(plotProgressBar())) plotProgressBar()$close()
+    plotProgressFile(NULL)
+    plotProgressBar(NULL)
+    # Hide/show and disable/enable relevant inputs
     shinyjs::hide("cancelPlot")
     shinyjs::show("doPlot")
     enableShinyInputs(input, except="^multiPlots*")